Fix loading of Validation Rule Schema from classpath
Out of the box, the gateway is configured to load the
validation rule schema from within the classpath,
rather than from a file on disk. This was resulting
in a failure[1] as `ResourceUtils.getFile()` does not
support loading from a classpath.

By switching to use `getURL().openStream()`, both
classpath and filesystem loads will work correctly.

[1]: `Caused by: org.springframework.beans.factory.BeanCreationException: Error creating bean with name 'validationRuleSchemaProvider': Invocation of init method failed; nested exception is java.io.FileNotFoundException: class path resource [validation-rule.schema.json] cannot be resolved to absolute file path because it does not reside in the file system: jar:file:/dgcg.jar!/BOOT-INF/classes!/validation-rule.schema.json`